this is false and even harmful
Chasing perfection has been scientifically associated with
procrastination, and because of this, it leads to undone tasks, not to
excellence. Chasing and catching excellence has surprisingly little to
do with perfectionism. See e.g. Fuschia M. Sirois, Procrastination -
What It Is, Why Its a Problem, and What You Can Do about, 2022.